LPS-28266

+++ portal-impl/src/com/liferay/portal/upgrade/v5_1_5/util/MBDiscussionDependencyManager.java
@@ -72,7 +72,7 @@ public class MBDiscussionDependencyManager extends DependencyManager {
 			rs = ps.executeQuery();
 
 			while (rs.next()) {
-				long count = rs.getLong(1);
+				int count = rs.getInt(1);
 
 				if (count > 0) {
 					return false;
