+++ portal-service/src/com/liferay/portal/kernel/upgrade/UpgradeProcess.java
@@ -185,7 +185,7 @@ public abstract class UpgradeProcess {
 			rs = ps.executeQuery();
 
 			while (rs.next()) {
-				long count = rs.getLong(1);
+				int count = rs.getInt(1);
 
 				if (count > 0) {
 					return true;
